# src/definitions/swagger.yaml
basePath: /
definitions:
...
    IMatch:
        description: ""
        properties:
            id:
                type: string
                description: ""
            bettingLeague:
                type: string
                description: ""
            gameNumber:
                type: string
                description: ""
            dateTime:
                type: string
                format: date-time
                description: ""
            homeTeam:
                type: string
                description: ""
            awayTeam:
                type: string
                description: ""
            homeScore:
                type: string
                description: ""
            awayScore:
                type: string
                description: ""
            overtime:
                type: boolean
                description: ""
            shotout:
                type: boolean
                description: ""
            winner:
                type: boolean
                description: ""
            isEvaluated:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - bettingLeague
            - gameNumber
            - dateTime
            - homeTeam
            - awayTeam
            - homeScore
            - awayScore
            - overtime
            - shotout
            - winner
            - isEvaluated
    IBettingLeague:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            sport:
                type: string
                description: ""
            isActive:
                type: boolean
                description: ""
            isTheMostActive:
                type: boolean
                description: ""
            seasonFrom:
                type: string
                description: ""
            seasonTo:
                type: string
                description: ""
            isFinished:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - name
            - sport
            - isActive
            - isTheMostActive
            - seasonFrom
            - seasonTo
            - isFinished
    IMatchScorer:
        description: ""
        properties:
            matchId:
                type: string
                description: ""
            scorerId:
                type: string
                description: ""
            numberOfGoals:
                type: string
                description: ""
        type: object
        required:
            - matchId
            - scorerId
            - numberOfGoals
    IPlayer:
        description: ""
        properties:
            id:
                type: string
                description: ""
            firstName:
                type: string
                description: ""
            lastName:
                type: string
                description: ""
            team:
                type: string
                description: ""
            seasonFrom:
                type: string
                description: ""
            seasonTo:
                type: string
                description: ""
            bestScorer:
                type: string
                description: ""
            isActive:
                type: boolean
                description: ""
            seasonGames:
                type: string
                description: ""
            seasonGoals:
                type: string
                description: ""
            seasonAssists:
                type: string
                description: ""
        type: object
        required:
            - id
            - firstName
            - lastName
            - team
            - seasonFrom
            - seasonTo
            - bestScorer
            - isActive
            - seasonGames
            - seasonGoals
            - seasonAssists
...
    ISpecialBetResult:
        description: ""
        properties:
            id:
                type: string
                description: ""
            bettingLeagueId:
                type: string
                description: ""
            specialBetId:
                type: string
                description: ""
            seriesHomeTeam:
                type: string
                description: ""
            seriesAwayTeam:
                type: string
                description: ""
            specialBetResult:
                type: string
                description: ""
        type: object
        required:
            - id
            - bettingLeagueId
            - specialBetId
            - seriesHomeTeam
            - seriesAwayTeam
            - specialBetResult
...
info:
    license:
        name: MIT
    title: tipapp-api
    version: 0.0.1
paths:
...
    /api/matches:
        get:
            operationId: MachesGetMatches
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IMatch'
            description: ""
            parameters: []
        post:
            operationId: MachesCreateMatch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IMatch'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: match
                    required: true
                    schema:
                        type: object
    '/api/matches/{id}':
        get:
            operationId: MachesGetMatch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IMatch'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: MachesUpdateMatch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IMatch'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: match
                    required: true
                    schema:
                        type: object
        delete:
            operationId: MachesDeleteMatch
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/betLeagues:
        get:
            operationId: BetLeaguesGetBetLeagues
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IBettingLeague'
            description: ""
            parameters: []
        post:
            operationId: BetLeaguesCreateBetLeague
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IBettingLeague'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: bet
                    required: true
                    schema:
                        type: object
    '/api/betLeagues/{id}':
        get:
            operationId: BetLeaguesGetBetLeague
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IBettingLeague'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: BetLeaguesUpdateBetLeague
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IBettingLeague'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: bet
                    required: true
                    schema:
                        type: object
        delete:
            operationId: BetLeaguesDeleteBetLeague
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/matchScorers:
        get:
            operationId: MatchScoresGetMatchScorers
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IMatchScorer'
            description: ""
            parameters: []
        post:
            operationId: MatchScoresCreateMatchScorer
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IMatchScorer'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: matchScorer
                    required: true
                    schema:
                        type: object
    '/api/matchScorers/{id}':
        get:
            operationId: MatchScoresGetMatchScorer
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IMatchScorer'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: MatchScoresUpdateMatchScorer
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IMatchScorer'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: matchScorer
                    required: true
                    schema:
                        type: object
        delete:
            operationId: MatchScoresDeleteMatchScorer
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/players:
        get:
            operationId: PlayersGetPlayers
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IPlayer'
            description: ""
            parameters: []
        post:
            operationId: PlayersCreatePlayer
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IPlayer'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: player
                    required: true
                    schema:
                        type: object
    '/api/players/{id}':
        get:
            operationId: PlayersGetPlayer
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IPlayer'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: PlayersUpdatePlayer
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IPlayer'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: player
                    required: true
                    schema:
                        type: object
        delete:
            operationId: PlayersDeletePlayer
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
...
    /api/specBetResults:
        get:
            operationId: SpecialBetResultsGetSpecialBetResults
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/ISpecialBetResult'
            description: ""
            parameters: []
        post:
            operationId: SpecialBetResultsCreateSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISpecialBetResult'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: specialBetResult
                    required: true
                    schema:
                        type: object
    '/api/specBetResults/{id}':
        get:
            operationId: SpecialBetResultsGetSpecialBetResult
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISpecialBetResult'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        put:
            operationId: SpecialBetResultsUpdateSpecialBet
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ISpecialBetResult'
            description: ""
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
                -
                    description: ""
                    in: body
                    name: specialBetResult
                    required: true
                    schema:
                        type: object
        delete:
            operationId: SpecialBetResultsDeleteSpecialBet
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
...
swagger: '2.0'
securityDefinitions: {}